#50 refactor: StoresProvider を _stores/index.tsx へ移動
- _contexts/stores-provider.tsx を _stores/index.tsx へ移動、各 store への import を相対パス短縮
- 呼び出し元 (time-control-03/index.tsx) の import 先を _stores へ変更
