+++ src/prototypes/time-control/time-control-03/index.tsx
@@ -3,7 +3,7 @@ import { ActionLogPanel } from './_components/action-log-panel'
 import { ActorController } from './_components/actor-controller'
 import { SchedulePreview } from './_components/schedule-preview'
 import { StageView } from './_components/stage-view'
-import { StoresProvider } from './_contexts/stores-provider'
+import { StoresProvider } from './_stores'
 import { ActorId } from './types'
 
 type TimeControl03Props = {
